SI units (#10)

+++ src/size_format.rs
@@ -39,7 +39,7 @@ const MAX_POWOF10: u64 = 1000;
 
 /**
 This trait adds [`format_size`](FormatSize::format_size) method
-to primitive [`u64`](core::u64) and [`usize`](core::u64) types.
+to primitive [`u64`](core::u64) and [`usize`](core::usize) types.
 */
 pub trait FormatSize {
     /// Splits the original size into integral, fractional and adds a unit.
@@ -83,7 +83,7 @@ impl FormatSize for Size {
 
 const UNITS: [&str; 7] = ["B", "KiB", "MiB", "GiB", "TiB", "PiB", "EiB"];
 
-#[cfg(all(test, not(feature = "no_std")))]
+#[cfg(all(test, feature = "std"))]
 mod tests {
     #![allow(clippy::panic)]
     use arbitrary::Arbitrary;
